Extract BaseSession request ID increment logic
This allows the request ID increment logic to be overwritten by
subclasses. I have a unique need to share sessions' underlying input
and output streams, but not the sessions themselves. This would require
keeping track of request IDs globally to avoid overlap.

Having this logic overridable lets me do that.